+++ pypicloud/__init__.py
@@ -4,14 +4,14 @@ import datetime
 import logging
 from pyramid.config import Configurator
 from pyramid.renderers import JSON, render
-from pyramid.settings import asbool, aslist
+from pyramid.settings import asbool
 from pyramid_beaker import session_factory_from_settings
 from six.moves.urllib.parse import urlencode  # pylint: disable=F0401,E0611
 
-from .cache import get_cache_impl
 from .route import Root
 
-__version__ = '0.2.1'
+
+__version__ = '0.2.2'
 LOG = logging.getLogger(__name__)
 
 
@@ -43,6 +43,7 @@ def includeme(config):
     config.include('pyramid_duh.auth')
     config.include('pypicloud.auth')
     config.include('pypicloud.access')
+    config.include('pypicloud.cache')
     settings = config.get_settings()
 
     config.add_renderer('json', json_renderer)
@@ -64,23 +65,17 @@ def includeme(config):
     config.registry.fallback_url = settings.get('pypi.fallback_url',
                                                 default_url)
 
-    fallback_modes = aslist(settings.get('pypi.fallback', ['redirect']))
+    fallback_mode = settings.get('pypi.fallback', 'redirect')
     # Compatibility with the deprecated pypi.use_fallback option
     if 'pypi.fallback' not in settings and 'pypi.use_fallback' in settings:
         LOG.warn("Using deprecated option 'pypi.use_fallback'")
         use_fallback = asbool(settings['pypi.use_fallback'])
-        fallback_modes = ['redirect'] if use_fallback else []
+        fallback_mode = 'redirect' if use_fallback else 'none'
     modes = ('redirect', 'cache', 'none')
-    for mode in fallback_modes:
-        if mode not in modes:
-            raise ValueError("Invalid value for 'pypi.fallback'. "
-                             "Must be one of %s" % ', '.join(modes))
-    config.registry.fallback = fallback_modes
-
-    # CACHING DATABASE SETTINGS
-    cache_impl = get_cache_impl(settings)
-
-    config.add_request_method(cache_impl, name='db', reify=True)
+    if fallback_mode not in modes:
+        raise ValueError("Invalid value for 'pypi.fallback'. "
+                         "Must be one of %s" % ', '.join(modes))
+    config.registry.fallback = fallback_mode
 
     # Special request methods
     config.add_request_method(_app_url, name='app_url')