+++ pypicloud/views/simple.py
@@ -1,7 +1,6 @@
 """ Views for simple pip interaction """
 import six
-from pyramid.httpexceptions import (HTTPBadRequest, HTTPFound, HTTPForbidden,
-                                    HTTPNotFound)
+from pyramid.httpexceptions import HTTPBadRequest, HTTPFound, HTTPNotFound
 from pyramid.view import view_config
 
 import posixpath
@@ -10,10 +9,9 @@ from pypicloud.util import normalize_name, FilenameScrapingLocator
 from pyramid_duh import argify, addslash
 
 
-@view_config(context=Root, request_method='POST', subpath=(),
-             permission='login', renderer='json')
+@view_config(context=Root, request_method='POST', subpath=(), renderer='json')
 @view_config(context=SimpleResource, request_method='POST', subpath=(),
-             permission='login', renderer='json')
+             renderer='json')
 @argify
 def upload(request, name, version, content):
     """ Handle update commands """
@@ -21,7 +19,7 @@ def upload(request, name, version, content):
     name = normalize_name(name)
     if action == 'file_upload':
         if not request.access.has_permission(name, 'write'):
-            return HTTPForbidden()
+            return request.forbid()
         try:
             return request.db.upload(content.filename, content.file, name=name,
                                      version=version)
@@ -55,28 +53,30 @@ def package_versions(context, request):
     normalized_name = normalize_name(context.name)
 
     packages = request.db.all(normalized_name)
-    if not packages:
-        if ('cache' in request.registry.fallback and
-                request.access.can_update_cache()):
-            locator = FilenameScrapingLocator(request.registry.fallback_url)
-            dists = locator.get_project(context.name)
-            if not dists:
-                return HTTPNotFound()
-            pkgs = {}
-            for dist in six.itervalues(dists):
-                filename = posixpath.basename(dist.source_url)
-                url = request.app_url('api', 'package', dist.name, filename)
-                pkgs[filename] = url
-            return {'pkgs': pkgs}
-        elif 'redirect' in request.registry.fallback:
-            redirect_url = "%s/%s/" % (
-                request.registry.fallback_url.rstrip('/'), context.name)
-            return HTTPFound(location=redirect_url)
-        else:
+    if packages:
+        if not request.access.has_permission(normalized_name, 'read'):
+            return request.forbid()
+        pkgs = {}
+        for package in packages:
+            pkgs[package.filename] = package.get_url(request)
+        return {'pkgs': pkgs}
+
+    elif request.registry.fallback == 'cache':
+        if not request.access.can_update_cache():
+            return request.forbid()
+        locator = FilenameScrapingLocator(request.registry.fallback_url)
+        dists = locator.get_project(context.name)
+        if not dists:
             return HTTPNotFound()
-    if not request.access.has_permission(normalized_name, 'read'):
-        raise HTTPForbidden()
-    pkgs = {}
-    for package in packages:
-        pkgs[package.filename] = package.get_url(request)
-    return {'pkgs': pkgs}
+        pkgs = {}
+        for dist in six.itervalues(dists):
+            filename = posixpath.basename(dist.source_url)
+            url = request.app_url('api', 'package', dist.name, filename)
+            pkgs[filename] = url
+        return {'pkgs': pkgs}
+    elif request.registry.fallback == 'redirect':
+        redirect_url = "%s/%s/" % (
+            request.registry.fallback_url.rstrip('/'), context.name)
+        return HTTPFound(location=redirect_url)
+    else:
+        return HTTPNotFound()