+++ tests/test_auth.py
@@ -55,6 +55,17 @@ class TestBasicAuth(MockServerTest):
         creds = auth.get_basicauth_credentials(self.request)
         self.assertEqual(creds, {'login': username, 'password': password})
 
+    def test_forbid(self):
+        """ When not logged in, forbid() returns 401 """
+        ret = auth._forbid(self.request)
+        self.assertEqual(ret.status_code, 401)
+
+    def test_forbid_logged_in(self):
+        """ When logged in, forbid() returns 403 """
+        self.request.userid = 'abc'
+        ret = auth._forbid(self.request)
+        self.assertEqual(ret.status_code, 403)
+
 
 class TestBasicAuthPolicy(MockServerTest):
 
@@ -93,21 +104,12 @@ class TestBasicAuthPolicy(MockServerTest):
 
     def test_principals_userid_no_credentials(self):
         """ Only [Everyone] if no credentials """
-        self.get_creds.return_value = None
-        principals = self.policy.effective_principals(self.request)
-        self.assertItemsEqual(principals, [Everyone])
-
-    def test_principals_fail_verification(self):
-        """ Only [Everyone] if access backend fails verification """
-        self.get_creds.return_value = {'login': 'dsa', 'password': 'foobar'}
-        self.request.access.verify_user.return_value = False
         principals = self.policy.effective_principals(self.request)
         self.assertItemsEqual(principals, [Everyone])
 
     def test_principals(self):
-        """ Return principals from access if basic auth succeeds """
-        self.get_creds.return_value = {'login': 'dsa', 'password': 'foobar'}
-        self.request.access.verify_user.return_value = True
+        """ Return principals from access if auth succeeds """
+        self.request.userid = 'dsa'
         principals = self.policy.effective_principals(self.request)
         self.request.access.user_principals.assert_called_with('dsa')
         self.assertEqual(principals, self.request.access.user_principals())
@@ -145,6 +147,10 @@ class TestSessionAuthPolicy(MockServerTest):
         self.request.session['user'] = 'dsa'
         userid = self.policy.authenticated_userid(self.request)
         self.assertEqual(userid, 'dsa')
+
+    def test_unauth_userid(self):
+        """ Unauth userid is pulled from request """
+        self.request.userid = 'dsa'
         userid = self.policy.unauthenticated_userid(self.request)
         self.assertEqual(userid, 'dsa')
 
@@ -154,8 +160,8 @@ class TestSessionAuthPolicy(MockServerTest):
         self.assertItemsEqual(principals, [Everyone])
 
     def test_principals(self):
-        """ Return principals from access if basic auth succeeds """
-        self.request.session['user'] = 'dsa'
+        """ Return principals from access if auth succeeds """
+        self.request.userid = 'dsa'
         principals = self.policy.effective_principals(self.request)
         self.request.access.user_principals.assert_called_with('dsa')
         self.assertEqual(principals, self.request.access.user_principals())